Better extract semantics; slight change to Pickett.json file format.
Extraction is now smart enough to handle dealing with objects that are
just random files/directories inside an image (although it creates a
container to do this work).  As before, you can refer to objects that
have been built into the source tree.  The result of both of these
operations, as needed, is merged into a tarball which is sent to the
docker server for a build.  This means the "build" procedure is called
only once no matter what mix of files you have from the soure tree
and some random image.

The OOD checking has been updated to include checking if artifacts
are up to date with the images that they will be placed in, which never
before worked.